Generate dynamic getters dynamically
Hardcoding a list of property names for which to generate getters works, but
doesn't scale well: it means that whenever Normaliz adds new properties, we
must adjust `PyNormaliz`; and if PyNormaliz then is used with an older version
of Normaliz, then there are wrappers that don't actually work. Similar if a
feature is simply disabled in Normaliz.

Thus, now instead generate those getters completely dynamically, by asking
Normaliz for a list of names of supported properties.